requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
certifi==2023.7.22
click==8.1.6
colorama==0.4.6
dnspython==2.4.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
idna==3.4
mako==1.2.4
markupsafe==2.1.3
passlib==1.7.4
pyasn1==0.5.0
pydantic-core==2.3.0
pydantic-extra-types==2.0.0
pydantic-partial==0.5.1
pydantic-settings==2.0.2
pydantic==2.0.3
pydantic[email]==2.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyyaml==6.0.1
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlalchemy==2.0.19
starlette==0.27.0
typing-extensions==4.7.1
uvicorn[standard]==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
